queenB-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.10.6
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
pydantic[email]>=2.7,<3.0